stratifiedresamples eval still to do, but: evalutators hierarchy updated a bit. all (aside from sr) record estimation timings and build timings correctly/separately. singletestset eval now records/infers meta info. an extra abstract class is inserted into hierarchy to denote that the evalautors will sample the data passed in some way and the classifier does not need to be prebuilt
