+++ src/main/java/evaluation/evaluators/CrossValidationEvaluator.java
@@ -38,7 +38,7 @@ import weka.core.Instances;
  * 
  * @author James Large (james.large@uea.ac.uk)
  */
-public class CrossValidationEvaluator extends SamplingEvaluator {
+public class CrossValidationEvaluator extends MultiSamplingEvaluator {
       
     //cursed code to allow tuning of regressors, should be removed if we ever delve deeper into regression stuff
     public static boolean REGRESSION_HACK = false;
